Add github workflow for continuous integration (#31)
1. Multi-platform (windows, macOS, ubuntu) and multi-compiler (clang++,
g++, and cl) build
2. Testing and coverage analysis (likely broken at this point - need
more work on build issues)
3. Formatting using clang-format
4. aws-labs/git-secrets to avoid pushing git credentials

Triggered only on pull request to main.

+++ nucleus/src/deployment/device_configuration.hpp
@@ -144,7 +144,7 @@ namespace deployment {
         const DeviceConfigConsts configs;
         static constexpr uint64_t COMPONENT_STORE_MAX_SIZE_DEFAULT_BYTES = 10'000'000'000L;
         static constexpr long DEPLOYMENT_POLLING_FREQUENCY_DEFAULT_SECONDS = 15L;
-        static constexpr uint32_t GG_DATA_PLANE_PORT_DEFAULT = 8443;
+        static constexpr uint64_t GG_DATA_PLANE_PORT_DEFAULT = 8443;
 
         DeviceConfiguration(data::Environment &environment, lifecycle::Kernel &kernel);
         DeviceConfiguration(const DeviceConfiguration &) = delete;
